feat: ✨ Added login and signup
In this commit, we have introduced the login and sign up features into the application. The existing auth functionality was enhanced to include Firebase authentication. The login functionality includes fields validation and password masking, whereas the sign up functionality includes fields validation, password and confirm password masking. ReviewText was moved and planned to be included in the gameView. The database version was updated to 2 and UserDao functions were made suspended. A check was performed to ascertain whether the input in the email field is valid. Minor changes have been made to the Signup.kt and the User.kt files for data compatibility to and from the Firebase. Commented out the permission request related code lines in Signup.kt and SignupViewModel.kt, thus dismissing the feature that relied on permissions. Replaced LiveData with MutableState due to Compose being the main UI toolkit.