refactor: ♻️ Refactor context activity lookup into utility class
The function findActivity() that was previously located within the SignupViewModel has now been moved to a separate utility class for better organisation and reusability. This function, that finds the activity context from a given context, is not solely tied to the scope of the SignupViewModel and thus was segregated to an external utility class.