chore: Update versionCode in build.gradle.kts
The versionCode in the build.gradle.kts file was incremented from 1 to 3. This change was necessary to prepare the app for a new release, as the versionCode denotes the version of the application and is used for update purposes. Incrementing this number signals that a newer version of the app is available.