fix: Fix top bar back navigation issue
Changed the condition for enabling the back navigation button in the main activity top bar. Instead of checking if the current destination is not the home destination, it now checks if the current destination is not any of the NavBarDestinations. This is done to accommodate all scenarios where the back button shouldn't be enabled, improving the user navigation experience.