feat: Refactor ProfileViewModel and add new features
This commit refactors the ProfileViewModel in the game library application to add new features,
including retrieval of followers and followed users. It also includes methods for getUser and
getLibrary to retrieve the profile's user and his/her library.
A new 'Follow' feature is added that required creation of 'FollowRepository' and 'FollowDao'.
User attributes in the 'User' data model are updated to include 'image' and 'bio' fields.
Methods to set image, username and bio have been added to UserDao. Various changes are made to
'Home' destination, including integration of user reviews and post components.
This commit will enhance profile-related features in our app, providing a better user experience.